Preview Item
Created an object with less API info to not overload the API with requests when whanting to show lots of items on screen.

This preview object serves for Media and Character API Objects. The intention is to store the API ids to later on request a larger info of the item when whanting to display it on detail.